+++ src/activitys/SoloShuffle.ts
@@ -15,8 +15,6 @@ import { VideoCategory } from '../types/VideoCategory';
  * Class representing a Solo Shuffle. This is essentially a wrapper around
  * a list of ArenaMatch objects, where the winner of the nested ArenaMatch
  * objects are determined by whoever gets the first kill.
- *
- * @@@ TODO handle leaver players (i.e. self), might just need a test?
  */
 export default class SoloShuffle extends Activity {
   private rounds: ArenaMatch[] = [];