+++ src/parsing/RetailLogHandler.ts
@@ -1,3 +1,8 @@
+import {
+  IArenaMatch,
+  IMalformedCombatData,
+  IShuffleMatch,
+} from 'wow-combat-log-parser';
 import Combatant from '../main/Combatant';
 
 import {
@@ -50,9 +55,6 @@ export default class RetailLogHandler extends LogHandler {
       .on('ARENA_MATCH_START', async (line: LogLine) => {
         await this.handleArenaStartLine(line);
       })
-      .on('ARENA_MATCH_END', async (line: LogLine) => {
-        await this.handleArenaEndLine(line);
-      })
       .on('CHALLENGE_MODE_START', async (line: LogLine) => {
         await this.handleChallengeModeStartLine(line);
       })
@@ -64,6 +66,15 @@ export default class RetailLogHandler extends LogHandler {
       })
       .on('SPELL_CAST_SUCCESS', async (line: LogLine) => {
         this.handleSpellCastSuccess(line);
+      })
+      .on('WAL_ARENA_END', async (event: IArenaMatch) => {
+        await this.handleWALArenaEnd(event);
+      })
+      .on('WAL_SHUFFLE_END', async (event: IShuffleMatch) => {
+        await this.handleWALShuffleMatchEnd(event);
+      })
+      .on('WAL_MALFORMED', async (event: IMalformedCombatData) => {
+        await this.handleWALParserMalformed(event);
       });
   }
 
@@ -481,5 +492,31 @@ export default class RetailLogHandler extends LogHandler {
     const endTime = line.date();
     this.activity.end(endTime, false);
     await this.endRecording();
+  }
+
+  private async handleWALArenaEnd(event: IArenaMatch) {
+    console.info('[RetailLogHandler] Got IArenaMatch event from WAL');
+
+    // The last line from the WAL event is the ARENA_MATCH_END line.
+    const endLine = new LogLine(event.rawLines[event.rawLines.length - 1]);
+    await this.handleArenaEndLine(endLine);
+  }
+
+  private async handleWALShuffleMatchEnd(event: IShuffleMatch) {
+    console.info('[RetailLogHandler] Got IShuffleMatch event from WAL');
+
+    // The last line from the WAL event is the ARENA_MATCH_END line.
+    const lastRound = event.rounds[5];
+    const endLine = new LogLine(
+      lastRound.rawLines[lastRound.rawLines.length - 1]
+    );
+
+    await this.handleArenaEndLine(endLine);
+  }
+
+  private async handleWALParserMalformed(event: IMalformedCombatData) {
+    console.warn('[RetailLogHandler] Got IMalformedCombatData event from WAL');
+    console.warn(event);
+    await this.forceEndActivity();
   }
 }